+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## main
 
+- Add `Connection#sid`, a unique connection identifier.
+
 - Add `Connection#broadcast` as an interface for broadcasting from channels.
 
 - Add `config.fastlane_broadcasts_enabled` to opt-in for optimized broadcasts (no double JSON encoding).